float32 before softmax

+++ models.py
@@ -71,7 +71,8 @@ class ContrastModel(keras.Model):
                 pred_logits, _ = self(imgs1)
 
             # Classifer cross entropy
-            unreduced_ce_loss = losses.sparse_categorical_crossentropy(labels, pred_logits, from_logits=True)
+            unreduced_ce_loss = losses.sparse_categorical_crossentropy(labels, tf.cast(pred_logits, tf.float32),
+                                                                       from_logits=True)
             ce_loss = tf.nn.compute_average_loss(unreduced_ce_loss, global_batch_size=bsz)
             loss = con_loss + ce_loss
